LPS-75840 SF - Rename class
