+++ modules/apps/adaptive-media/adaptive-media-image-api/src/main/java/com/liferay/adaptive/media/image/scaler/AMImageScaler.java
@@ -67,7 +67,7 @@ public interface AMImageScaler {
 	 *
 	 * @review
 	 */
-	public AMImageScaled scaleImage(
+	public AMImageScaledImage scaleImage(
 		FileVersion fileVersion,
 		AMImageConfigurationEntry amImageConfigurationEntry);
